Asynchronous Postgres calls implemented

+++ app/db/models/announcement_db.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Boolean, Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import BigInteger, Boolean, Column, Integer, String, ForeignKey, DateTime
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 from app.db.models.dungeon_run_db import DungeonRunDB
@@ -12,8 +12,8 @@ class AnnouncementDB(Base):
     id = Column(Integer, primary_key=True)
     title = Column(String)
     content = Column(String)
-    discord_guild_id = Column(Integer)
-    announcement_channel_id = Column(Integer)
+    discord_guild_id = Column(BigInteger)
+    announcement_channel_id = Column(BigInteger)
     has_been_sent = Column(Boolean)
     created_at = Column(DateTime, default=datetime.utcnow)
     modified_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
