+++ app/db/models/character_history_db.py
@@ -1,5 +1,5 @@
 import datetime
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Boolean
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String, Boolean
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 
@@ -9,8 +9,8 @@ class CharacterHistoryDB(Base):
     __tablename__ = 'character_history'
     
     id = Column(Integer, primary_key=True)
-    discord_user_id = Column(Integer, nullable=False)
-    discord_guild_id = Column(Integer, nullable=False)    
+    discord_user_id = Column(BigInteger, nullable=False)
+    discord_guild_id = Column(BigInteger, nullable=False)    
     guild_name = Column(String, nullable=False)    
     name = Column(String, nullable=False)
     realm = Column(String, nullable=False)
